requirements.txt
TwitterAPI==2.5.4
python-dotenv-0.8.2
selenium-3.13.0
